+++ tests/config/phpstan-custom-app-config.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+use yii2\extensions\phpstan\tests\support\stub\ApplicationCustom;
+
+return [
+    'phpstan' => [
+        'application_type' => ApplicationCustom::class,
+    ],
+];